fix: Improve gca to handle new repositories
Previously, the `gca` command would error if the local branch did not have a corresponding remote branch. This would happen in new repositories where the remote branch has not yet been pushed.

This commit modifies the `gca` command to only check for the remote branch if there are one or more commits in the local branch. This allows `gca` to be used in new repositories without error.

(This commit message was automatically generated by an LLM.)