feat: Add BitWarden API client secret to autoinstall
This commit adds the BitWarden API client secret to the autoinstall configuration. This allows the post-installation script to automatically log in to BitWarden without prompting the user for their password. The client secret is written to disk temporarily and then deleted by the post-installation script.

The following changes were made:
- Added `apikey` to the `cspell.config.jsonc` words list.
- Added `CLIENTID` and `CLIENTSECRET` to the `cspell.config.jsonc` words list.
- Added a command to write the BitWarden API client secret to disk in the `ubuntu-auto-install/autoinstall.yaml` file.
- Modified the `ubuntu-auto-install/post-install/post-install.sh` script to use the BitWarden API client secret for login.
- Removed the master password from the `ubuntu-auto-install/autoinstall.yaml` file, as it is no longer needed.
- Updated the BitWarden login command to use the `--apikey` flag.

(This commit message was automatically generated by an LLM.)