fix(bash): Improve os-release sourcing robustness
The change modifies the condition for sourcing `/etc/os-release` from `[ -s /etc/os-release ]` to `[[ -s "/etc/os-release" ]]`. This is a minor improvement for robustness by using bash's conditional expression `[[ ... ]]` which is generally preferred for string and file tests over `[ ... ]`. It also quotes the path `/etc/os-release` which is good practice.

(This commit message was automatically generated by an LLM.)