feat: add get_num_commits_on_branch function
This function calculates the number of commits on the current branch since its divergence from the main or master branch.

It also updates the `gpr` function to use the new `get_num_commits_on_branch` function. If there is only one commit on the branch, it uses the description of that commit as the pull request title. Otherwise, it uses the LLM to generate the pull request text.

(This commit message was automatically generated by an LLM.)